프로토콜 추가 및 클라이언트 내 서버와의 통신 추가
클라이언트도 서버와 같은 WSAAsyncSelect 모델을 사용하여 서버에게서 메시지를 받아
현재 진행 되는 씬에서 패킷을 Recv하여 각 씬을 해당 패킷의 정보로 갱신할 것 이다.

+++ AvoidHimClient/Timer.cpp
@@ -19,8 +19,8 @@ Timer::~Timer() {
 
 }
 
-double Timer::GetTimeElapsed() const {
-	return timeElapsed;
+float Timer::GetTimeElapsed() const {
+	return (float)timeElapsed;
 }
 
 int Timer::GetFPS() const {
